requirements.txt
discord.py
transformers
py-dotenv
torch
psycopg2-binary